Move can UI into App.tsx with other UI elements
