+++ src/App.tsx
@@ -12,12 +12,15 @@ require('codemirror/mode/clike/clike');
 
 import WorkerInstance from './WorkerInstance';
 import { RobotState } from './RobotState';
+import Collapsible from './Collapsible';
 
 
 export interface AppProps extends StyleProps {
   robot: RobotState;
+  isCanChecked: boolean[];
 
   onRobotChange: (robot: RobotState) => void;
+  onCanChange: (canNumber: number, checked: boolean) => void;
 }
 
 interface AppState {
@@ -161,6 +164,12 @@ export class App extends React.Component<Props, State> {
     });
   };
 
+  private onCheckBoxActivity = (event: React.ChangeEvent<HTMLInputElement>) => {
+		const canNumber = Number.parseInt(event.target.value);
+    const isTargetChecked = event.target.checked;
+    this.props.onCanChange(canNumber, isTargetChecked);
+	}
+
   render() {
     //let codemirror = new CodeMirrorManager();
     const { props, state } = this;
@@ -175,84 +184,96 @@ export class App extends React.Component<Props, State> {
     }
     //console.log("Rendering app");
     return (
-      <section className="app-area">
-        <section className="app-header">
-          <svg width="197" height="164" className="logo">
-            <image height="100%" href="static/KIPR-Logo-bk.jpg"/>
-          </svg>
-          <h1 className="ide-title">KISS IDE<br/>Simulator</h1>
-        </section>
-        <p>
-          <button onClick={this.onCompileClick_}>Compile</button>
-          <button onClick={this.onRunClick_}>Run</button>
-          <button onClick={this.onDownloadClick_}>Download</button>
-        </p>
-        <CodeMirror rows={20} cols={65} value={code} onChange={this.onCodeChange_} options={options} id="code" name="code" className="code" />
-        <textarea rows={10} cols={70} value={console} contentEditable={false} />
-        <section className="robotState">
-          <h3 className="robotStateHead">Robot State</h3>
-          <button onClick={this.onResetClick_} className="resetButton">Reset</button>
-          <section className="robotStateValues">
-            <section className="x">
-              <h3>X: </h3>
-              <textarea rows={1} cols={5} draggable="false" onChange={this.onXChange_} value={Math.round(robot.x)}/>
+      <React.Fragment>
+        <section className="app-area">
+          <section className="app-header">
+            <svg width="197" height="164" className="logo">
+              <image height="100%" href="static/KIPR-Logo-bk.jpg"/>
+            </svg>
+            <h1 className="ide-title">KISS IDE<br/>Simulator</h1>
+          </section>
+          <p>
+            <button onClick={this.onCompileClick_}>Compile</button>
+            <button onClick={this.onRunClick_}>Run</button>
+            <button onClick={this.onDownloadClick_}>Download</button>
+          </p>
+          <CodeMirror rows={20} cols={65} value={code} onChange={this.onCodeChange_} options={options} id="code" name="code" className="code" />
+          <textarea rows={10} cols={70} value={console} contentEditable={false} />
+          <section className="robotState">
+            <h3 className="robotStateHead">Robot State</h3>
+            <button onClick={this.onResetClick_} className="resetButton">Reset</button>
+            <section className="robotStateValues">
+              <section className="x">
+                <h3>X: </h3>
+                <textarea rows={1} cols={5} draggable="false" onChange={this.onXChange_} value={Math.round(robot.x)}/>
+              </section>
+              <section className="y">
+                <h3>Y: </h3>
+                <textarea rows={1} cols={5} draggable="false" onChange={this.onYChange_} value={Math.round(robot.y)}/>
+              </section>
+              <section className="theta">
+                <h3>Theta: </h3>
+                <textarea rows={1} cols={5} draggable="false" onChange={this.onThetaChange_} value={Math.round(-1*robot.theta/Math.PI*180)}/>
+              </section>
             </section>
-            <section className="y">
-              <h3>Y: </h3>
-              <textarea rows={1} cols={5} draggable="false" onChange={this.onYChange_} value={Math.round(robot.y)}/>
+          </section>
+          <section className="motors">
+            <section className="motorValues">
+              <h3>Motor:</h3>
+              <h3>Speed:</h3>
+              <h3>Position:</h3>
             </section>
-            <section className="theta">
-              <h3>Theta: </h3>
-              <textarea rows={1} cols={5} draggable="false" onChange={this.onThetaChange_} value={Math.round(-1*robot.theta/Math.PI*180)}/>
+            <section className="motor0">
+              <h3>Motor 0</h3>
+              <div><textarea rows={1} cols={4} draggable="false" readOnly value={Math.round(robot.motor0_speed)} /></div>
+              <div><textarea rows={1} cols={4} draggable="false" readOnly value={Math.round(robot.motor0_position)} /></div>
+            </section>
+            <section className="motor1">
+              <h3>Motor 1</h3>
+              <div><textarea rows={1} cols={4} draggable="false" readOnly value={Math.round(robot.motor1_speed)} /></div>
+              <div><textarea rows={1} cols={4} draggable="false" readOnly value={Math.round(robot.motor1_position)} /></div>
+            </section>
+            <section className="motor2">
+              <h3>Motor 2</h3>
+              <div><textarea rows={1} cols={4} draggable="false" readOnly value={Math.round(robot.motor2_speed)} /></div>
+              <div><textarea rows={1} cols={4} draggable="false" readOnly value={Math.round(robot.motor2_position)} /></div>
+            </section>
+            <section className="motor3">
+              <h3>Motor 3</h3>
+              <div><textarea rows={1} cols={4} draggable="false" readOnly value={Math.round(robot.motor3_speed)} /></div>
+              <div><textarea rows={1} cols={4} draggable="false" readOnly value={Math.round(robot.motor3_position)} /></div>
+            </section>
+          </section>
+          <section className="servos">
+            <section className="servo0">
+              <h3>Servo 0</h3>
+              <textarea rows={1} cols={4} draggable="false" readOnly /*onChange={this.onServo0Change_}*/ value={Math.round(robot.servo0_position)} />
+            </section>
+            <section className="servo1">
+              <h3>Servo 1</h3>
+              <textarea rows={1} cols={4} draggable="false" readOnly /*onChange={this.onServo1Change_}*/ value={Math.round(robot.servo1_position)} />
+            </section>
+            <section className="servo2">
+              <h3>Servo 2</h3>
+              <textarea rows={1} cols={4} draggable="false" readOnly /*onChange={this.onServo2Change_}*/ value={Math.round(robot.servo2_position)} />
+            </section>
+            <section className="servo3">
+              <h3>Servo 3</h3>
+              <textarea rows={1} cols={4} draggable="false" readOnly /*onChange={this.onServo3Change_}*/ value={Math.round(robot.servo3_position)} />
             </section>
           </section>
         </section>
-        <section className="motors">
-          <section className="motorValues">
-            <h3>Motor:</h3>
-            <h3>Speed:</h3>
-            <h3>Position:</h3>
-          </section>
-          <section className="motor0">
-            <h3>Motor 0</h3>
-            <div><textarea rows={1} cols={4} draggable="false" readOnly value={Math.round(robot.motor0_speed)} /></div>
-            <div><textarea rows={1} cols={4} draggable="false" readOnly value={Math.round(robot.motor0_position)} /></div>
-          </section>
-          <section className="motor1">
-            <h3>Motor 1</h3>
-            <div><textarea rows={1} cols={4} draggable="false" readOnly value={Math.round(robot.motor1_speed)} /></div>
-            <div><textarea rows={1} cols={4} draggable="false" readOnly value={Math.round(robot.motor1_position)} /></div>
-          </section>
-          <section className="motor2">
-            <h3>Motor 2</h3>
-            <div><textarea rows={1} cols={4} draggable="false" readOnly value={Math.round(robot.motor2_speed)} /></div>
-            <div><textarea rows={1} cols={4} draggable="false" readOnly value={Math.round(robot.motor2_position)} /></div>
-          </section>
-          <section className="motor3">
-            <h3>Motor 3</h3>
-            <div><textarea rows={1} cols={4} draggable="false" readOnly value={Math.round(robot.motor3_speed)} /></div>
-            <div><textarea rows={1} cols={4} draggable="false" readOnly value={Math.round(robot.motor3_position)} /></div>
-          </section>
-        </section>
-        <section className="servos">
-          <section className="servo0">
-            <h3>Servo 0</h3>
-            <textarea rows={1} cols={4} draggable="false" readOnly /*onChange={this.onServo0Change_}*/ value={Math.round(robot.servo0_position)} />
-          </section>
-          <section className="servo1">
-            <h3>Servo 1</h3>
-            <textarea rows={1} cols={4} draggable="false" readOnly /*onChange={this.onServo1Change_}*/ value={Math.round(robot.servo1_position)} />
-          </section>
-          <section className="servo2">
-            <h3>Servo 2</h3>
-            <textarea rows={1} cols={4} draggable="false" readOnly /*onChange={this.onServo2Change_}*/ value={Math.round(robot.servo2_position)} />
-          </section>
-          <section className="servo3">
-            <h3>Servo 3</h3>
-            <textarea rows={1} cols={4} draggable="false" readOnly /*onChange={this.onServo3Change_}*/ value={Math.round(robot.servo3_position)} />
-          </section>
-        </section>
-      </section>
+        <Collapsible title="Cans">
+          <ul>
+            {[...Array(12)].map((_, i) =>
+              <li key={i + 1}>
+                <input type="checkbox" name="can" value={i} checked={this.props.isCanChecked[i]} onChange={this.onCheckBoxActivity} />
+                <label>{`Can ${i + 1}`}</label>
+              </li>
+            )}
+          </ul>
+        </Collapsible>
+      </React.Fragment>
     )
   }
 }
